Add files via upload
This is a project by Jayden Claytor and I for submission in the Washington State University 2025 Hackathon. It is a program that is designed to share the beauty of fractals, both visually and mathematically. It was certainly a challenge to program visuals for 9 different fractals; we had to learn some complex analysis, how to use raylib.h for graphics, and what a fractal truly is in its mathematical form.